Refactor feature computation module
