+++ requirements.txt
@@ -1,11 +1,14 @@
-e3fp>=1.2.3
+# rdkit-pypi>=2021.9.5.1
+rdkit-pypi==2021.3.1.1
+# e3fp>=1.2.3
 mordred==1.2.0
 networkx==2.3.0
 numpy>=1.21.2
 scipy>=1.7.3
 padelpy>=0.1.11
-pandas>=1.4.0
+pandas>=1.3.5
 pytest>=6.2.5
+coverage>=6.3.2
 pytest-cov>=3.0.0
 setuptools>=58.0.4
-rdkit>=2020.09.1.0
+sklearn>=1.0.0
