+++ tox.ini
@@ -10,16 +10,13 @@ python =
 envlist = py37, py38, py39, rst_linux, rst_mac, readme, rst, linters, coverage-report, qa
 
 [testenv]
+; conda_deps =
+;     rdkit
+; conda_channels = rdkit
 deps =
-    pytest
-    pytest-cov
-    coverage>=4.2
-    numpy
-    scipy
-    pandas
-    # sphinx
+    -r{toxinidir}/requirements.txt
 commands =
-# coverage run --rcfile=tox.ini -m pytest tests
+    # coverage run --rcfile=tox.ini -m pytest tests
     python -m pip install --upgrade pip
     pip install .
 # coverage run --source=DiverseSelector --rcfile=tox.ini -m pytest DiverseSelector DiverseSelector/test
